Refactoring
Split up each component, assigned states as props, made an interface file, tried to tackle some bugs.
