+++ HM/Pages/Navigation/Views/HabitScreen/HabitScreen.tsx
@@ -1,24 +1,11 @@
 import * as React from "react";
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  Button,
-  Modal,
-  TextInput,
-  FlatList,
-} from "react-native";
+import { View, Text, Button, FlatList } from "react-native";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import uuid from "react-native-uuid";
-import { styles, modalViewStyles } from "./styles";
-import { Weekdays, getCurrentDay, randomHexColor } from "../../../utils/utils";
-
-interface HabitData {
-  id: string;
-  taskName: string;
-  selectedDays: string[];
-  color: string;
-}
+import { styles } from "./styles";
+import { Weekdays, getCurrentDay } from "../../../utils/utils";
+import ModalView from "./components/ModalView";
+import HabitItem from "./components/ListofHabits";
+import { HabitData } from "../../../Interfaces/interfaces";
 
 const HomeScreen: React.FC = ({}) => {
   const currentDay = getCurrentDay();
@@ -35,51 +22,6 @@ const HomeScreen: React.FC = ({}) => {
         day.id === selectedID ? { ...day, isSelected: !day.isSelected } : day
       )
     );
-  };
-
-  const createTask = async () => {
-    if (!taskName.trim()) {
-      console.error("Task name cannot be empty");
-      return;
-    }
-
-    const selectedDays = dayList
-      .filter((day) => day.isSelected)
-      .map((day) => day.name);
-
-    const taskData = {
-      key: uuid.v4(),
-      taskName,
-      selectedDays,
-      color: randomHexColor(),
-    };
-
-    try {
-      const savedHabitsString = await AsyncStorage.getItem("habits");
-      const currentHabits = savedHabitsString
-        ? JSON.parse(savedHabitsString)
-        : [];
-
-      const updatedHabits = [
-        ...currentHabits,
-        {
-          key: taskData.key,
-          taskName: taskData.taskName,
-          selectedDays: taskData.selectedDays,
-          color: taskData.color,
-        },
-      ];
-
-      await AsyncStorage.setItem("habits", JSON.stringify(updatedHabits));
-
-      setHabits(updatedHabits.map((habit) => habit.taskName));
-
-      fetchHabits();
-    } catch (error) {
-      console.error("Error saving task", error);
-    }
-
-    setModalVisible(false);
   };
 
   const fetchHabits = async () => {
@@ -171,140 +113,28 @@ const HomeScreen: React.FC = ({}) => {
 
   return (
     <View style={styles.viewContainer}>
-      <Modal
-        animationType="fade"
-        transparent={true}
-        visible={modalVisible}
-        onRequestClose={() => {
-          setModalVisible(false);
-        }}
-      >
-        <TouchableOpacity
-          style={styles.overlay}
-          activeOpacity={1}
-          onPress={() => setModalVisible(false)}
-        >
-          <View style={modalViewStyles.centeredView}>
-            <View style={modalViewStyles.modalView}>
-              <View style={{ width: "80%", alignContent: "center" }}>
-                <TextInput
-                  style={modalViewStyles.modalText}
-                  onChangeText={setTaskName}
-                  placeholder="Habit name"
-                />
-              </View>
-              <Text style={modalViewStyles.dueDays}>Due days: </Text>
-              <View
-                style={{
-                  flexDirection: "row",
-                  alignContent: "center",
-                }}
-              >
-                {dayList.map((day) => (
-                  <TouchableOpacity
-                    key={day.id}
-                    style={{
-                      backgroundColor: day.isSelected ? "#5669FF" : "darkgray",
-                      padding: 10,
-                      width: "10%",
-                      alignItems: "center",
-                      borderRadius: 5,
-                      margin: 5,
-                    }}
-                    onPress={() => handlePress(day.id)}
-                  >
-                    <Text style={{ color: "white" }}>{day.name.charAt(0)}</Text>
-                  </TouchableOpacity>
-                ))}
-              </View>
-              <TouchableOpacity
-                style={[modalViewStyles.button, modalViewStyles.buttonClose]}
-                onPress={createTask}
-              >
-                <Text style={modalViewStyles.textStyle}>Create task</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-        </TouchableOpacity>
-      </Modal>
-      <Text
-        style={{
-          fontSize: 26,
-          fontWeight: "bold",
-          color: modalVisible ? "black" : "white",
-        }}
-      >
-        Habits
-      </Text>
-
+      <ModalView
+        modalVisible={modalVisible}
+        setModalVisible={setModalVisible}
+        taskName={taskName}
+        setTaskName={setTaskName}
+        dayList={dayList}
+        handlePress={handlePress}
+        setHabits={setHabits}
+        setSelectedDaysForHabit={setSelectedDaysForHabit}
+      />
+      <Text style={{ fontSize: 26, fontWeight: "bold" }}>Habits</Text>
       <FlatList
         data={habits}
         renderItem={({ item }) => (
-          <View style={styles.habitContainer}>
-            <Text style={styles.habitText}>{item}</Text>
-            <View style={styles.buttonsContainer}>
-              {Weekdays.map((day) => (
-                <TouchableOpacity
-                  key={day.id}
-                  style={[
-                    styles.dayButton,
-                    selectedDaysForHabit[item] &&
-                    selectedDaysForHabit[item].includes(day.name)
-                      ? styles.dayButtonSelected
-                      : styles.dayButtonDisabled,
-                  ]}
-                  onPress={() => {
-                    if (!selectedDaysForHabit[item]) {
-                      handleCheckIn(item, day.name);
-                    }
-                  }}
-                  disabled={
-                    !selectedDaysForHabit[item] ||
-                    !selectedDaysForHabit[item].includes(day.name)
-                  }
-                >
-                  <Text
-                    style={[
-                      styles.dayButtonText,
-                      selectedDaysForHabit[item] &&
-                      selectedDaysForHabit[item].includes(day.name)
-                        ? styles.dayButtonTextSelected
-                        : selectedDaysForHabit[item] &&
-                          !selectedDaysForHabit[item].includes(day.name)
-                        ? styles.dayButtonTextDisabled
-                        : null,
-                      checkedInDays[item] &&
-                      checkedInDays[item].includes(day.name)
-                        ? styles.dayButtonTextCheckedIn
-                        : null,
-                    ]}
-                    onPress={() => {
-                      console.log(
-                        "selectedDaysForHabit[item]: ",
-                        selectedDaysForHabit[item]
-                      );
-                    }}
-                  >
-                    {day.name.charAt(0)}
-                  </Text>
-                </TouchableOpacity>
-              ))}
-            </View>
-            <View style={styles.habitButtons}>
-              <TouchableOpacity
-                style={styles.checkInButton}
-                onPress={() => handleCheckIn(item, currentDay)}
-              >
-                <Text style={styles.checkInButtonText}>Check-in</Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                style={styles.removeButton}
-                onPress={() => handleRemoveHabit(item)}
-              >
-                <Text style={styles.removeButtonText}>Remove</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
+          <HabitItem
+            item={item}
+            selectedDaysForHabit={selectedDaysForHabit}
+            checkedInDays={checkedInDays}
+            handleCheckIn={handleCheckIn}
+            handleRemoveHabit={handleRemoveHabit}
+            currentDay={currentDay}
+          />
         )}
         keyExtractor={(item) => item}
       />
